refactor: decompose selector system into parse/match/resolve/build modules
Closes #159
